# [REVIEW] Use CuPy v8 FFT cache plan