--- a/python/cusignal/convolution/convolve.py
+++ b/python/cusignal/convolution/convolve.py
@@ -303,39 +303,55 @@ def fftconvolve(in1, in2, mode="full", axes=None):
     fshape = [next_fast_len(d) for d in shape[axes]]
     fslice = tuple([slice(sz) for sz in shape])
 
-    if not complex_result:
-        if (str(fshape), str(axes), "R2C") in _cupy_fft_cache:
-            rplan = _cupy_fft_cache[(str(fshape), str(axes), "R2C")]
-        else:
-            rplan = _cupy_fft_cache[
-                (str(fshape), str(axes), "R2C")
-            ] = fftpack.get_fft_plan(in1, fshape, axes=axes, value_type="R2C")
-        try:
-            with rplan:
-                sp1 = cp.fft.rfftn(in1, fshape, axes=axes)
-                sp2 = cp.fft.rfftn(in2, fshape, axes=axes)
-        except Exception:
+    ver = cp.__version__
+    print(ver)
+    print(ver.split('.'))
+    major_ver = cp.__version__.split('.')
+    
+    if int(major_ver[0]) >= 8:
+        if not complex_result:
             sp1 = cp.fft.rfftn(in1, fshape, axes=axes)
             sp2 = cp.fft.rfftn(in2, fshape, axes=axes)
-
-        ret = cp.fft.irfftn(sp1 * sp2, fshape, axes=axes)[fslice].copy()
-    else:
-        # Need to move to cupyx.scipy.fft with CuPy v8
-        if (str(fshape), str(axes)) in _cupy_fft_cache:
-            plan = _cupy_fft_cache[(str(fshape), str(axes))]
+            ret = cp.fft.irfftn(sp1 * sp2, fshape, axes=axes)[fslice].copy()
         else:
-            plan = _cupy_fft_cache[
-                (str(fshape), str(axes))
-            ] = fftpack.get_fft_plan(in1, fshape, axes=axes)
-        try:
-            with plan:
-                sp1 = fftpack.fftn(in1, fshape, axes=axes)
-                sp2 = fftpack.fftn(in2, fshape, axes=axes)
-                ret = fftpack.ifftn(sp1 * sp2, axes=axes)[fslice].copy()
-        except Exception:
             sp1 = fftpack.fftn(in1, fshape, axes=axes)
             sp2 = fftpack.fftn(in2, fshape, axes=axes)
             ret = fftpack.ifftn(sp1 * sp2, axes=axes)[fslice].copy()
+    else: # CuPy v7
+        if not complex_result:
+            if (str(fshape), str(axes), "R2C") in _cupy_fft_cache:
+                rplan = _cupy_fft_cache[(str(fshape), str(axes), "R2C")]
+            else:
+                rplan = _cupy_fft_cache[
+                    (str(fshape), str(axes), "R2C")
+                ] = fftpack.get_fft_plan(in1, fshape, axes=axes, value_type="R2C")
+            try:
+                with rplan:
+                    sp1 = cp.fft.rfftn(in1, fshape, axes=axes)
+                    sp2 = cp.fft.rfftn(in2, fshape, axes=axes)
+            except Exception:
+                sp1 = cp.fft.rfftn(in1, fshape, axes=axes)
+                sp2 = cp.fft.rfftn(in2, fshape, axes=axes)
+
+            ret = cp.fft.irfftn(sp1 * sp2, fshape, axes=axes)[fslice].copy()
+        else:
+            # Need to move to cupyx.scipy.fft with CuPy v8
+            if (str(fshape), str(axes)) in _cupy_fft_cache:
+                plan = _cupy_fft_cache[(str(fshape), str(axes))]
+            else:
+                plan = _cupy_fft_cache[
+                    (str(fshape), str(axes))
+                ] = fftpack.get_fft_plan(in1, fshape, axes=axes)
+            try:
+                with plan:
+                    sp1 = fftpack.fftn(in1, fshape, axes=axes)
+                    sp2 = fftpack.fftn(in2, fshape, axes=axes)
+                    ret = fftpack.ifftn(sp1 * sp2, axes=axes)[fslice].copy()
+            except Exception:
+                sp1 = fftpack.fftn(in1, fshape, axes=axes)
+                sp2 = fftpack.fftn(in2, fshape, axes=axes)
+                ret = fftpack.ifftn(sp1 * sp2, axes=axes)[fslice].copy()
+
 
     if mode == "full":
         return ret