@mnicely 2020-10-05
> Hi @mnicely, thanks for the benchmark data. I am a bit confused --- any chance the outcomes for v8 + cache and v8 no cache are swapped? The cached performance seems to be worse if I read it right.

Not swapped, just bad wording! That is the cuSignal cache I created. When I turn it off and use CuPy’s, the FFT is faster!