@leofang 2020-10-05
Ahh I see, thanks for clarifying, Matt. The number looks very good then! I wonder if all can be attributed to CuPy's cache, or there are additional nice changes made to v8?