@mnicely 2020-10-05
I believe the speedups between cuSignal’s cache + CuPy v7.8 and cuSignal (with no cache) + CuPy v8.0 is solely the cache. 

And the differences cuSignal (with no cache) between v7.8 and v8.0 is attributed to many improvements.