@mnicely 2020-10-05
> Thanks, @mnicely! I wonder if you could do one additional test for me when you have time: Use CuPy v8, but turn off all caches (either cuSignal's or CuPy's). The latter can be turned off this way:
> 
> ```python
> import cupy as cp
> 
> cache = cp.fft.config.get_plan_cache()
> cache.set_size(0)
> ```
> 
> Note the `cache` object is per thread & per device, so if your tests span over threads and/or devices, you need to turn them all off in the proper context. For confirmation, if you do `cache.show_info()`, you'd see a line saying `cache enabled? False`.

Sure, I'll try to have you something by the end of this week!